+++ js/app/packages/core/component/LexicalMarkdown/plugins/actions/actions.ts
@@ -15,8 +15,7 @@ import TextH3 from '@icon/regular/text-h-three.svg';
 import TextH2 from '@icon/regular/text-h-two.svg';
 import TextT from '@icon/regular/text-t.svg';
 import { INSERT_TABLE_COMMAND, TableNode } from '@lexical/table';
-import type { KlassConstructor, LexicalEditor, LexicalNode } from 'lexical';
-import type { Component } from 'solid-js';
+import type { LexicalEditor } from 'lexical';
 import { INSERT_HORIZONTAL_RULE_COMMAND } from '..';
 import { TRY_INSERT_EQUATION_COMMAND } from '../katex';
 import { TRY_INSERT_LINK_COMMAND } from '../links';
@@ -36,28 +35,7 @@ import { ListNode } from '@lexical/list';
 import { LinkNode } from '@lexical/link';
 import { globalSplitManager } from '@app/signal/splitLayout';
 import type { ComposeTaskSuccess } from '@block-md/component/ComposeTask';
-
-export type ActionIcon = string;
-
-export type Action = {
-  id: string;
-  name: string;
-  keywords: string[];
-  icon: Component<{ class: string }>;
-  category: string;
-  action: (editor: LexicalEditor) => void;
-  shortcut?: string;
-  dependencies?: Array<KlassConstructor<typeof LexicalNode>>;
-};
-
-// TODO (seamus): Actually organize the items based on category.
-export enum ActionCategory {
-  BASIC = 'Basic',
-  FORMAT = 'Formatting',
-  ELEMENT = 'Elements',
-  MEDIA = 'Media',
-  ADVANCED = 'Advanced',
-}
+import { type Action, ActionCategory } from './types';
 
 export const ACTIONS: Action[] = [
   {
