+++ js/app/packages/core/component/LexicalMarkdown/plugins/actions/index.ts
@@ -1,4 +1,5 @@
 export * from './actions';
+export * from './types';
 export {
   actionsPlugin,
   CLOSE_ACTION_SEARCH_COMMAND,
